+++ libsolidity/analysis/DeclarationContainer.h
@@ -24,9 +24,9 @@
 #pragma once
 
 #include <libsolidity/ast/ASTForward.h>
+#include <liblangutil/Exceptions.h>
+#include <liblangutil/SourceLocation.h>
 #include <boost/noncopyable.hpp>
-#include <map>
-#include <set>
 
 namespace solidity::frontend
 {
@@ -38,17 +38,26 @@ namespace solidity::frontend
 class DeclarationContainer
 {
 public:
+	using Homonyms = std::vector<std::pair<langutil::SourceLocation const*, std::vector<Declaration const*>>>;
+
 	explicit DeclarationContainer(
 		ASTNode const* _enclosingNode = nullptr,
-		DeclarationContainer const* _enclosingContainer = nullptr
+		DeclarationContainer* _enclosingContainer = nullptr
 	):
-		m_enclosingNode(_enclosingNode), m_enclosingContainer(_enclosingContainer) {}
+		m_enclosingNode(_enclosingNode), m_enclosingContainer(_enclosingContainer)
+	{
+		if (_enclosingContainer)
+			_enclosingContainer->m_innerContainers.emplace_back(this);
+	}
 	/// Registers the declaration in the scope unless its name is already declared or the name is empty.
 	/// @param _name the name to register, if nullptr the intrinsic name of @a _declaration is used.
-	/// @param _invisible if true, registers the declaration, reports name clashes but does not return it in @a resolveName
-	/// @param _update if true, replaces a potential declaration that is already present
+	/// @param _location alternative location, used to point at homonymous declarations.
+	/// @param _invisible if true, registers the declaration, reports name clashes but does not return it in @a resolveName.
+	/// @param _update if true, replaces a potential declaration that is already present.
 	/// @returns false if the name was already declared.
-	bool registerDeclaration(Declaration const& _declaration, ASTString const* _name = nullptr, bool _invisible = false, bool _update = false);
+	bool registerDeclaration(Declaration const& _declaration, ASTString const* _name, langutil::SourceLocation const* _location, bool _invisible, bool _update);
+	bool registerDeclaration(Declaration const& _declaration, bool _invisible, bool _update);
+
 	std::vector<Declaration const*> resolveName(ASTString const& _name, bool _recursive = false, bool _alsoInvisible = false) const;
 	ASTNode const* enclosingNode() const { return m_enclosingNode; }
 	DeclarationContainer const* enclosingContainer() const { return m_enclosingContainer; }
@@ -67,11 +76,18 @@ public:
 	/// Searches this and all parent containers.
 	std::vector<ASTString> similarNames(ASTString const& _name) const;
 
+	/// Populates a vector of (location, declaration) pairs, where location is a location of an inner-scope declaration,
+	/// and declaration is the corresponding homonymous outer-scope declaration.
+	void populateHomonyms(std::back_insert_iterator<Homonyms> _it) const;
+
 private:
 	ASTNode const* m_enclosingNode;
 	DeclarationContainer const* m_enclosingContainer;
+	std::vector<DeclarationContainer const*> m_innerContainers;
 	std::map<ASTString, std::vector<Declaration const*>> m_declarations;
 	std::map<ASTString, std::vector<Declaration const*>> m_invisibleDeclarations;
+	/// List of declarations (name and location) to check later for homonymity.
+	std::vector<std::pair<std::string, langutil::SourceLocation const*>> m_homonymCandidates;
 };
 
 }