+++ libsolidity/formal/BMC.cpp
@@ -396,7 +396,6 @@ void BMC::endVisit(FunctionCall const& _funCall)
 	case FunctionType::Kind::Send:
 	case FunctionType::Kind::Transfer:
 	{
-		SMTEncoder::endVisit(_funCall);
 		auto value = _funCall.arguments().front();
 		solAssert(value, "");
 		smtutil::Expression thisBalance = m_context.state().balance();
@@ -406,6 +405,8 @@ void BMC::endVisit(FunctionCall const& _funCall)
 			thisBalance < expr(*value),
 			&_funCall
 		);
+
+		SMTEncoder::endVisit(_funCall);
 		break;
 	}
 	case FunctionType::Kind::AddMod: