+++ libsolidity/formal/PredicateSort.cpp
@@ -28,49 +28,54 @@ using namespace solidity::smtutil;
 namespace solidity::frontend::smt
 {
 
-SortPointer interfaceSort(ContractDefinition const& _contract)
+SortPointer interfaceSort(ContractDefinition const& _contract, SymbolicState& _state)
 {
 	return make_shared<FunctionSort>(
-		stateSorts(_contract),
+		vector<SortPointer>{_state.thisAddressSort(), _state.stateSort()} + stateSorts(_contract),
 		SortProvider::boolSort
 	);
 }
 
-SortPointer nondetInterfaceSort(ContractDefinition const& _contract)
+SortPointer nondetInterfaceSort(ContractDefinition const& _contract, SymbolicState& _state)
 {
 	auto varSorts = stateSorts(_contract);
+	vector<SortPointer> stateSort{_state.stateSort()};
 	return make_shared<FunctionSort>(
-		varSorts + varSorts,
+		stateSort + varSorts + stateSort + varSorts,
 		SortProvider::boolSort
 	);
 }
 
-SortPointer implicitConstructorSort()
+SortPointer implicitConstructorSort(SymbolicState& _state)
 {
-	return arity0FunctionSort();
+	return make_shared<FunctionSort>(
+		vector<SortPointer>{_state.errorFlagSort(), _state.thisAddressSort(), _state.stateSort()},
+		SortProvider::boolSort
+	);
 }
 
-SortPointer constructorSort(ContractDefinition const& _contract)
+SortPointer constructorSort(ContractDefinition const& _contract, SymbolicState& _state)
 {
 	if (auto const* constructor = _contract.constructor())
-		return functionSort(*constructor, &_contract);
+		return functionSort(*constructor, &_contract, _state);
 
 	return make_shared<FunctionSort>(
-		vector<SortPointer>{SortProvider::uintSort} + stateSorts(_contract),
+		vector<SortPointer>{_state.errorFlagSort(), _state.thisAddressSort(), _state.stateSort()} + stateSorts(_contract),
 		SortProvider::boolSort
 	);
 }
 
-SortPointer functionSort(FunctionDefinition const& _function, ContractDefinition const* _contract)
+SortPointer functionSort(FunctionDefinition const& _function, ContractDefinition const* _contract, SymbolicState& _state)
 {
 	auto smtSort = [](auto _var) { return smt::smtSortAbstractFunction(*_var->type()); };
 	auto varSorts = _contract ? stateSorts(*_contract) : vector<SortPointer>{};
 	auto inputSorts = applyMap(_function.parameters(), smtSort);
 	auto outputSorts = applyMap(_function.returnParameters(), smtSort);
 	return make_shared<FunctionSort>(
-		vector<SortPointer>{SortProvider::uintSort} +
+		vector<SortPointer>{_state.errorFlagSort(), _state.thisAddressSort(), _state.stateSort()} +
 			varSorts +
 			inputSorts +
+			vector<SortPointer>{_state.stateSort()} +
 			varSorts +
 			inputSorts +
 			outputSorts,
@@ -78,9 +83,9 @@ SortPointer functionSort(FunctionDefinition const& _function, ContractDefinition
 	);
 }
 
-SortPointer functionBodySort(FunctionDefinition const& _function, ContractDefinition const* _contract)
+SortPointer functionBodySort(FunctionDefinition const& _function, ContractDefinition const* _contract, SymbolicState& _state)
 {
-	auto fSort = dynamic_pointer_cast<FunctionSort>(functionSort(_function, _contract));
+	auto fSort = dynamic_pointer_cast<FunctionSort>(functionSort(_function, _contract, _state));
 	solAssert(fSort, "");
 
 	auto smtSort = [](auto _var) { return smt::smtSortAbstractFunction(*_var->type()); };