+++ test/libsolidity/smtCheckerTests/blockchain_state/this_does_not_change_external_call.sol
@@ -1,0 +1,18 @@
+pragma experimental SMTChecker;
+
+abstract contract D {
+	function d() external virtual;
+}
+
+contract C {
+	address t;
+	constructor() {
+		t = address(this);
+	}
+	function f(D d) public {
+		address a = address(this);
+		d.d();
+		assert(address(this) == t);
+		assert(a == t);
+	}
+}