+++ test/libsolidity/syntaxTests/abiEncoder/v1_inheritance_from_contract_calling_v2_function.sol
@@ -1,0 +1,25 @@
+==== Source: A ====
+pragma experimental ABIEncoderV2;
+
+struct Data {
+    bool flag;
+}
+
+contract A {
+    function get() public view returns (Data memory) {}
+}
+
+contract B {
+    constructor() {
+        A(0x00).get();
+    }
+
+    function foo() public view {
+        A(0x00).get();
+    }
+}
+==== Source: B ====
+import "A";
+
+contract C is B {}
+// ----