+++ test/libsolidity/syntaxTests/scoping/double_variable_declaration_disjoint_scope_activation.sol
@@ -5,5 +5,6 @@ contract test {
     }
 }
 // ----
+// Warning 2519: (57-63): This declaration shadows an existing declaration.
 // Warning 2072: (57-63): Unused local variable.
 // Warning 2072: (75-81): Unused local variable.